experimental settings

+++ ftrace/run2.sh
@@ -22,12 +22,13 @@ PING_ARGS="-D -i 1.0 -s 56"
 NATIVE_PING_CMD="${HOME}/contools-daemon/iputils/ping"
 CONTAINER_PING_CMD="/iputils/ping"
 
-PING_CONTAINER_IMAGE="ping-ubuntu"
+PING_CONTAINER_IMAGE="chrismisa/contools:ping-ubuntu"
 PING_CONTAINER_NAME="ping-container"
 
 PAUSE_CMD="sleep 5"
 
 PING_PAUSE_CMD="sleep 500"
+# PING_PAUSE_CMD="sleep 10"
 
 MONITOR_CMD="$(pwd)/latency $(pwd)/latency.conf"
 
@@ -35,8 +36,8 @@ DATE_TAG=`date +%Y%m%d%H%M%S`
 META_DATA="Metadata"
 
 #declare -a IPERF_ARGS=("1M" "3M" "10M" "32M" "100M" "316M" "1G" "3G" "10G")
-declare -a IPERF_ARGS=("1M" "10M" "100M" "1G" "10G")
-#declare -a IPERF_ARGS=("1M" "3M")
+declare -a IPERF_ARGS=("nop" "500K" "1M" "100M" "1G" "10G")
+#declare -a IPERF_ARGS=("nop" "3M")
 
 mkdir $DATE_TAG
 cd $DATE_TAG
@@ -59,9 +60,12 @@ $PAUSE_CMD
 for arg in ${IPERF_ARGS[@]}
 do
   # Start iperf
-  iperf -c $TARGET_IPV4 -d -i 100 -b $arg -t 0 \
-    > ${arg}.iperf &
-  IPERF_PID=$!
+  if [ $arg != "nop" ]
+  then
+    iperf -c $TARGET_IPV4 -i 100 -b $arg -t 0 \
+      > ${arg}.iperf &
+    IPERF_PID=$!
+  fi
   #
   # Native pings for control
   #
@@ -124,8 +128,11 @@ do
   
   $PAUSE_CMD
 
-  kill -INT $IPERF_PID
-  echo "  killed iperf"
+  if [ $arg != "nop" ]
+  then
+    kill -INT $IPERF_PID
+    echo "  killed iperf"
+  fi
 
   $PAUSE_CMD
